Split the two 950-line handoff files by responsibility
`handoff/mod.rs` (964 lines) and the phone PWA's `App.tsx` (947) had each
grown to hold several unrelated jobs.

The desktop side is now three modules: `mod.rs` keeps the endpoint
lifecycle and identity, `transfer.rs` owns one connection — token check,
op dispatch, and the wire helpers — and `transcribe.rs` owns the
transcribe op and its analytics event.

On the phone, `use-handoff-session.ts` takes the whole state machine
(pairing, capabilities, the durable outbox, recording, one send end to
end), leaving `App.tsx` as markup.

Behaviour is unchanged. Two small interface shifts fell out of the PWA
split: `onUnpair` no longer closes the settings sheet, since that is
`App`'s state, and the outbox delete is a `discardQueued` callback.
